requirements.txt
asgiref==3.7.2
asttokens==2.4.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.3
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.4
django-allauth==0.56.0
django-cors-headers==4.2.0
django-extensions==3.2.3
djangorestframework==3.14.0
exceptiongroup==1.2.2
executing==2.1.0
idna==3.4
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.7
oauthlib==3.2.2
parso==0.8.4
pillow==10.4.0
prompt_toolkit==3.0.47
pure_eval==0.2.3
pycparser==2.21
Pygments==2.18.0
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.13
